backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
insightface==0.7.3
onnxruntime==1.16.3
ultralytics==8.0.196
pandas==2.1.1
